+++ CHANGELOG.md
@@ -17,6 +17,7 @@ _Unreleased_
 - Fixed "unauthorized" error which occurred while updating email and password at
   same time.
 - Improve message for `machines list` when no machines are found.
+- When encrypting or signing, do not use revoked keypairs.
 
 ## v0.17.0